nom_rfc8288

Module complete

Structs

LinkData

The LinkData struct is used to store the URL provided in the Link header, as well as optional parameters.

LinkDataOwned

LinkParam

LinkParam is used represent the parsed data. It stores key value pairs from the Link header

LinkParamOwned

Enums

LinkParseError

ParseStrictness 🔒

Constants

NUM_EMPTY_ELEMENTS 🔒

Functions

linkDeprecated

Same as link_strict, but with the original name. When lenient parsing was introduced, to avoid ambiguity the function was renamed. To avoid breakages, the function still exists, but instead use link_strict.

link_inner 🔒

link_lenient

Same as link_strict, except that empty parameters are leniently parsed. They don’t result in an error and are skipped when collecting parameters.

link_strict

This method will parse a &str and return an array of Options if it can successfully parse the &str as a Link header. The reason we return Options is because if the Link header has empty elements, we want to show that information to the user by returning Nones.

list

qdtext

QDTEXT = HTAB / SP / %x21 / %x23-5B / %x5D-7E / obs-text

quoted_pair 🔒

QUOTED-PAIR = "\" ( HTAB / SP / VCHAR / obs-text )

quoted_string 🔒

QUOTED-STRING = DQUOTE *( qdtext / quoted-pair ) DQUOTE

quoted_string_alloca 🔒

tchar

token

TOKEN = 1*TCHAR